requirements.txt
beautifulsoup4==4.7.1
pycryptodome==3.9.4
requests==2.18.4
